remove mocked values from requests

+++ control_client/src/main/java/at/rmbt/client/control/Responses.kt
@@ -127,7 +127,7 @@ data class SettingItem(
 
 @Keep
 data class TermsAndConditionsSettings(
-    val version: Long?,
+    val version: Int?,
     val url: String?,
     @SerializedName("ndt_url")
     val ndtURL: String?
